requirements.txt
# Core dependencies
numpy>=1.19.0
pillow>=8.0.0
torch>=1.9.0
torchvision>=0.10.0

# 3D processing
open3d>=0.13.0
trimesh>=3.9.0
pyrender>=0.1.45

# Image processing
opencv-python>=4.5.0
scipy>=1.7.0

# Machine learning
huggingface-hub>=0.8.0
diffusers>=0.10.0
transformers>=4.20.0

# Utilities
tqdm>=4.62.0
requests>=2.26.0
matplotlib>=3.4.0

# Optional: For web interface
# fastapi>=0.68.0
# uvicorn>=0.15.0
# python-multipart>=0.0.5